Update the version of Python installed in the dailyaws job
Update from version pin from 3.10.2 to 3.11. The build is currently
failing because 3.10.2 is not available for Ubuntu 22.04 (the new
underlying version for ubuntu-latest tagged runners). Since we have to
update versions it makes sense to jump to the latest minor release of
Python 3. We do not pin to a patch release so that we can automatically
get the latest patch release available when the workflow is run. Since
this will include any improvements or security fixes without breaking
compatibility with the minor version this is generally a good practice.